+++ guillotina/api/app.py
@@ -8,8 +8,16 @@ from guillotina.interfaces import IResourceSerializeToJson
 
 @configure.service(
     context=IApplication, method='GET', permission='guillotina.AccessContent',
-    title="Get application data",
-    description="Retrieves serialization of application")
+    summary="Get application data",
+    description="Retrieves serialization of application",
+    responses={
+        "200": {
+            "description": "Application data",
+            "schema": {
+                "$ref": "#/definitions/Application"
+            }
+        }
+    })
 async def get(context, request):
     serializer = getMultiAdapter(
         (context, request),
@@ -18,9 +26,9 @@ async def get(context, request):
 
 
 @configure.service(
-    context=IApplication, method='GET', permission='guillotina.GetPortals',
-    name='@apidefinition',
-    title="Get API Definition",
+    context=IApplication, method='GET',
+    permission='guillotina.GetContainers', name='@apidefinition',
+    summary="Get API Definition",
     description="Retrieves information on API configuration")
 async def get_api_definition(context, request):
     return app_settings['api_definition']