+++ guillotina/api/behaviors.py
@@ -7,44 +7,77 @@ from guillotina.content import get_cached_factory
 from guillotina.interfaces import IBehavior
 from guillotina.interfaces import IResource
 from guillotina.interfaces import ISchemaSerializeToJson
+from guillotina.browser import Response
 
 
-@configure.service(context=IResource, method='PATCH', permission='guillotina.ModifyContent',
-                   name='@behaviors')
+@configure.service(
+    context=IResource, method='PATCH',
+    permission='guillotina.ModifyContent', name='@behaviors',
+    summary="Add behavior to resource",
+    parameters=[{
+        "name": "body",
+        "in": "body",
+        "schema": {
+            "$ref": "#/definitions/Behavior"
+        }
+    }],
+    responses={
+        "200": {
+            "description": "Successfully added behavior"
+        },
+        "201": {
+            "description": "Behavior already assigned here"
+        },
+    })
 async def default_patch(context, request):
-    """We add a behavior.
-
-    We expect on the body to be :
-    {
-        'behavior': 'INTERFACE.TO.BEHAVIOR.SCHEMA'
-    }
-    """
     data = await request.json()
     behavior = data.get('behavior', None)
+    if behavior in context.__behaviors__:
+        return Response(response={}, status=201)
     context.add_behavior(behavior)
     return {}
 
 
-@configure.service(context=IResource, method='DELETE', permission='guillotina.ModifyContent',
-                   name='@behaviors')
+@configure.service(
+    context=IResource, method='DELETE',
+    permission='guillotina.ModifyContent', name='@behaviors',
+    summary="Remove behavior from resource",
+    parameters=[{
+        "name": "body",
+        "in": "body",
+        "schema": {
+            "$ref": "#/definitions/Behavior"
+        }
+    }],
+    responses={
+        "200": {
+            "description": "Successfully removed behavior"
+        },
+        "201": {
+            "description": "Behavior not assigned here"
+        },
+    })
 async def default_delete(context, request):
-    """We add a behavior.
-
-    We expect on the body to be :
-    {
-        'behavior': 'INTERFACE.TO.BEHAVIOR.SCHEMA'
-    }
-    """
     data = await request.json()
     behavior = data.get('behavior', None)
+    if behavior not in context.__behaviors__:
+        return Response(response={}, status=201)
     context.remove_behavior(behavior)
     return {}
 
 
 @configure.service(
-    context=IResource, method='GET', permission='guillotina.AccessContent',
-    name='@behaviors',
-    description='Get information on behaviors for this resource')
+    context=IResource, method='GET',
+    permission='guillotina.AccessContent', name='@behaviors',
+    summary='Get information on behaviors for this resource',
+    responses={
+        "200": {
+            "description": "A listing of behaviors for content",
+            "schema": {
+                "$ref": "#/definitions/BehaviorsResponse"
+            }
+        }
+    })
 async def default_get(context, request):
     """We show the available schemas."""
     result = {}