+++ guillotina/testing.py
@@ -31,7 +31,8 @@ TESTING_SETTINGS = {
     "port": TESTING_PORT,
     "static": [
         {"static": os.path.dirname(os.path.realpath(__file__))},
-        {'favicon.ico': 'static/favicon.ico'}
+        {"module_static": 'guillotina:'},
+        {'favicon.ico': os.path.join(os.path.dirname(os.path.realpath(__file__)), '__init__.py')}
     ],
     "default_static_filenames": ['teststatic.txt'],
     "creator": {